auto

+++ detection/CD-session-KF/auto.sh
@@ -1,3 +1,9 @@
+if [ "$2" = "" ]
+then
+    echo "no argument: time ./auto.sh FILENAME 2017 10 15"
+    exit
+fi
+
 rm -rf kf_*
 
 rm -rf kf_all_in
@@ -13,10 +19,15 @@ while read line; do
     rm -rf kf_in_${line}
     #python kf2.py in_${line}_all instlist
     python kf4.py in_${line}_all instlist
+
+    python date-trans-2.py in_${line}_all $2 $3 $4 > in_${line}_all_dated
+    python date-trans-2.py kf_in_${line} $2 $3 $4 > kf_in_${line}_dated
+    
     cat kf_in_${line} >> kf_all_in
 done < $1
 
 ./sort.pl kf_all_in > tmp_in
+python date-trans-2.py tmp_in $2 $3 $4 > tmp_in_dated
 
 # output: timeslot, score,     ID,   name
 #              17, 0.00717402,XXXX, YYYY
@@ -25,8 +36,14 @@ while read line; do
     echo $line
     rm -rf kf_out_${line}
     python kf4.py out_${line}_all instlist
+
+    python date-trans-2.py out_${line}_all $2 $3 $4 > out_${line}_all_dated
+
+    python date-trans-2.py kf_out_${line} $2 $3 $4 > kf_out_${line}_dated
+
+
     cat kf_out_${line} >> kf_all_out
 done < $1
 
 ./sort.pl kf_all_out > tmp_out
-
+python date-trans-2.py tmp_out $2 $3 $4 > tmp_out_dated
